Combined _commits with _actions, and touched up some logging
